requirements.txt
discord==2.2.2
python-dotenv==1.0.0
revChatGPT==4.1.6
openai==0.27.4
asgiref==3.6.0